usdm: export dma_{map,unmap}_slab
These functions are required to add IOMMU mappings. If memory is
allocated using the qaeMemAlloc, this function is called internally, but
if using an external buffer, we had no way of adding a new mapping.

An issue has been opened upstream
(https://github.com/intel/qatlib/issues/31) but it until it has been
fixed, use this fork instead.